pyproject.toml
[tool.poetry]
name = "fastapi-basic"
version = "0.1.0"
description = ""
authors = ["Евгений <evgeniy.moiseev.progr@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
pydantic-settings = "^2.6.1"
pyjwt = "^2.10.1"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
pydantic = {version = "^2.10.2", extras = ["email"]}
passlib = "^1.7.4"
asyncpg = "^0.30.0"
user-agents = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
